pyproject.toml
[project]
name = "carousel"
version = "0.1.0"
description = "A Stable Marriage Solver."
readme = "README.md"
authors = [{ name = "Thomas (Tom) C. Gorordo", email = "tcgorordo@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "numpy>=2.2.4",
    "polars>=1.26.0",
    "pyside6>=6.9.0",
    "rich>=14.0.0",
]

[project.scripts]
carousel = "carousel:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "faker>=37.1.0",
    "hypothesis>=6.130.8",
    "marimo[recommended]>=0.13.6",
    "pyclean>=3.1.0",
    "pyinstaller>=6.12.0",
    "pyright>=1.1.398",
    "pytest>=8.3.5",
    "pytest-benchmark>=5.1.0",
    "ruff>=0.11.2",
    "ty>=0.0.0a5",
]
srv = [
    "legacy-cgi>=2.6.3",
    "python-dotenv>=1.1.0",
]
  
[pytest]
testpaths = "test"
log_cli = true

[tool.pyright]
include = ["src"]
exclude = ["test"]

reportMissingImports = "error"
reportMissingTypeStubs = false
